+++ src/client/src/commons/search/components/listFilter.js
@@ -5,6 +5,7 @@ import { Checkbox, Input, TextArea, Form } from "semantic-ui-react";
 import TranslationText from "../../form/translationText";
 import _ from "lodash";
 import { useTranslation } from "react-i18next";
+import HierarchicalDataSearch from "../../form/borehole/stratigraphy/hierarchicalDataSearch";
 import DomainDropdown from "../../form/domain/dropdown/domainDropdown";
 import DomainTree from "../../form/domain/tree/domainTree";
 import DateField from "../../form/dateField";
@@ -46,6 +47,9 @@ const ListFilter = props => {
     let isVisibleCounter = 0;
 
     for (let i = 0; i < attribute?.length; i++) {
+      if (attribute[i]?.hideShowAllFields === true) {
+        return false;
+      }
       if (isVisibleFunction(attribute[i]?.isVisibleValue)) {
         isVisibleCounter++;
       }
@@ -53,7 +57,9 @@ const ListFilter = props => {
 
     if (isVisibleCounter === attribute?.length) {
       return false;
-    } else return true;
+    } else {
+      return true;
+    }
   };
 
   const updateChange = (attribute, value, to = true, isNumber = false) => {
@@ -66,7 +72,8 @@ const ListFilter = props => {
       item.type === "Dropdown" ||
       item.type === "Canton" ||
       item.type === "City" ||
-      item.type === "ReferenceSystem"
+      item.type === "ReferenceSystem" ||
+      item.type === "HierarchicalData"
     ) {
       updateChange(item.value, null, false);
     } else if (item.type === "Radio") {
@@ -174,11 +181,12 @@ const ListFilter = props => {
               <Styled.AttributesContainer>
                 {(item.isVisible ||
                   isVisibleFunction(item.isVisibleValue) ||
-                  showAll) && (
-                  <Styled.Label>
-                    <TranslationText id={item.label} />
-                  </Styled.Label>
-                )}
+                  showAll) &&
+                  item.type !== "HierarchicalData" && (
+                    <Styled.Label>
+                      <TranslationText id={item.label} />
+                    </Styled.Label>
+                  )}
                 {item.type === "Input" &&
                   (item.isVisible ||
                     isVisibleFunction(item.isVisibleValue) ||
@@ -340,6 +348,24 @@ const ListFilter = props => {
                     </Styled.AttributesItem>
                   )}
 
+                {item.type === "HierarchicalData" &&
+                  (item.isVisible ||
+                    isVisibleFunction(item.isVisibleValue) ||
+                    showAll) && (
+                    <HierarchicalDataSearch
+                      onSelected={e => {
+                        updateChange(item.value, e.id, false);
+                      }}
+                      schema={item.schema}
+                      labels={item.labels}
+                      selected={
+                        _.isNil(search.filter?.[item.value])
+                          ? null
+                          : search.filter[item.value]
+                      }
+                    />
+                  )}
+
                 {item.type === "DomainTree" &&
                   (item.isVisible ||
                     isVisibleFunction(item.isVisibleValue) ||
@@ -395,6 +421,7 @@ const ListFilter = props => {
                 {(item.isVisible ||
                   isVisibleFunction(item.isVisibleValue) ||
                   showAll) &&
+                  item.type !== "HierarchicalData" &&
                   !item.hasTwoFields && (
                     <Styled.Reset>
                       <LabelReset
@@ -408,6 +435,7 @@ const ListFilter = props => {
                 {(item.isVisible ||
                   isVisibleFunction(item.isVisibleValue) ||
                   showAll) &&
+                  item.type !== "HierarchicalData" &&
                   item.hasTwoFields &&
                   item.label === "" && (
                     <Styled.Reset>