+++ src/client/src/commons/search/searchComponent.js
@@ -16,7 +16,7 @@ import { InstrumentSearchData } from "./data/InstrumentSearchData";
 import { fillingSearchData } from "./data/fillingSearchData";
 import { LocationSearchData } from "./data/LocationSearchData";
 import { boreholeSearchData } from "./data/boreholeSearchData";
-import { stratigraphySearchData } from "./data/stratigraphySearchData";
+import { lithologySearchData } from "./data/lithologySearchData";
 class SearchComponent extends React.Component {
   constructor(props) {
     super(props);
@@ -38,8 +38,8 @@ class SearchComponent extends React.Component {
         },
         {
           id: 2,
-          name: "stratigraphy",
-          translationId: "stratigraphy",
+          name: "lithology",
+          translationId: "lithology",
           isSelected: false,
         },
         {
@@ -104,10 +104,10 @@ class SearchComponent extends React.Component {
     ) {
       selectedData = boreholeSearchData;
     } else if (
-      this.state?.searchList?.[2]?.name === "stratigraphy" &&
+      this.state?.searchList?.[2]?.name === "lithology" &&
       this.state?.searchList?.[2]?.isSelected
     ) {
-      selectedData = stratigraphySearchData;
+      selectedData = lithologySearchData;
     } else if (
       this.state?.searchList?.[3]?.name === "casing" &&
       this.state?.searchList?.[3]?.isSelected